+++ src/main/java/com/greedy/mokkoji/common/log/query/ConnectionProxyHandler.java
@@ -1,4 +1,4 @@
-package com.greedy.mokkoji.common.log;
+package com.greedy.mokkoji.common.log.query;
 
 import lombok.RequiredArgsConstructor;
